refactor(gamemode): move server/client code to their modules
instead of inexor/gamemode/capture_server.hpp its inexor/server/gamemode/capture_server.hpp now.
this is since we can not define SharedVars in headers and the set of SharedVars is different for the server and the client.
