+++ inexor/client/gamemode/ctf_client.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "inexor/fpsgame/game.hpp"
 #include "inexor/gamemode/ctf_common.hpp"
-#include "inexor/gamemode/gamemode_client.hpp"
+#include "inexor/client/gamemode/gamemode_client.hpp"
 
 namespace game {
