+++ inexor/fpsgame/ai.cpp
@@ -1,5 +1,6 @@
 #include "inexor/fpsgame/game.hpp"
 #include "inexor/fpsgame/entities.hpp"
+#include "inexor/client/gamemode/gamemode_client.hpp"
 
 
 extern SharedVar<int> fog;
