--- inexor/fpsgame/bomb.hpp
@@ -1,4 +1,0 @@
-//NO INCLUDE GUARD
-// bomb.h: client and server state for bomb gamemode
-
-
