--- inexor/fpsgame/capture.hpp
@@ -1,6 +1,0 @@
-//NO INCLUDE GUARD
-// capture.h: client and server state for capture gamemode
-
-
-
-
