--- inexor/fpsgame/collect.hpp
@@ -1,2 +1,0 @@
-
-
