--- inexor/fpsgame/ctf.hpp
@@ -1,1 +1,0 @@
-
