--- inexor/fpsgame/hideandseek.hpp
@@ -1,3 +1,0 @@
-// bomb.h: client and server state for bomb gamemode
-
-
