+++ inexor/fpsgame/render.cpp
@@ -1,5 +1,6 @@
 #include "inexor/fpsgame/game.hpp"
 #include "inexor/fpsgame/entities.hpp"
+#include "inexor/client/gamemode/gamemode_client.hpp"
 
 using namespace inexor::sound;
